requirements.txt
streamlit==1.26.0
Pillow==10.0.0
cairosvg==2.7.0